backend/requirements.txt
# Core FastAPI and async dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Agno framework and tools
agno
mem0ai

# Appwrite SDK
appwrite==13.0.1

# Pydantic for data validation
pydantic==2.5.0

# Async HTTP client
httpx==0.25.2

# WebSocket support
websockets==12.0

# JSON handling
orjson==3.9.10

# Environment and configuration
python-decouple==3.8

# Logging
structlog==23.2.0

# Date/time handling
python-dateutil==2.8.2

# UUID generation
uuid==1.30

# Async utilities
asyncio-mqtt==0.16.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0 